new commit

+++ 0x0B-malloc_free/0-create_array.c
@@ -18,7 +18,7 @@ char *create_array(unsigned int size, char c)
 	}
 	else
 	{
-		for (i = 0; i < size; i++)
+		for (i = 0; i < (int) size; i++)
 		{
 			arptr[i] = c;
 		}
